Refactor OverlappedButtonsHelper
Make it so tooltips come out of the anchor element and not the icon,
which aligns it better on the reset guide button. In turn, padding
is replaced by margin in mu-overlapped so the tooltips are positioned
correctly in the actual overlapped buttons.

+++ app/helpers/overlapped_buttons_helper.rb
@@ -1,25 +1,35 @@
 module OverlappedButtonsHelper
   def expand_icon
-    overlapped_button_icon :fullscreen, :expand
-  end
-
-  def restart_icon(data_placement='left')
-    overlapped_button_icon :restart, :undo, data_placement
+    overlapped_link :fullscreen, :expand, class: 'editor-resize'
   end
 
   def format_icon
-    overlapped_button_icon :format, :indent
+    overlapped_link :format, :indent, class: 'editor-format'
+  end
+
+  def restart_icon
+    overlapped_link :restart, :undo, class: 'editor-reset submission-reset', 'data-confirm': t(:confirm_reset)
   end
 
   def restart_guide_link(guide)
-    link_to restart_icon('top'),
-            guide_progress_path(guide),
-            class: 'mu-content-toolbar-item mu-restart-guide',
-            data: {confirm: t(:confirm_restart)},
-            method: :delete
+    all_options = tooltip_options(:restart).merge!(
+      {class: 'mu-content-toolbar-item mu-restart-guide',
+       data: {confirm: t(:confirm_restart)}, method: :delete, 'data-bs-placement': 'top'})
+
+    link_to overlapped_button_icon(:undo), guide_progress_path(guide), all_options
   end
 
-  def overlapped_button_icon(key, icon, data_placement='left')
-    fa_icon(icon, title: t(key), class: 'fa-fw', role: 'button', 'aria-label': t(key), 'data-bs-placement': data_placement)
+  private
+
+  def overlapped_link(key, icon, **options)
+    content_tag :a, overlapped_button_icon(icon), tooltip_options(key).merge!(options)
+  end
+
+  def overlapped_button_icon(icon)
+    fa_icon icon, class: 'fa-fw', role: 'button'
+  end
+
+  def tooltip_options(key)
+    {title: t(key), 'aria-label': t(key), 'data-bs-toggle': 'tooltip', 'data-bs-placement': 'left'}
   end
 end
